Change Installation Instructions in README.md
The current instructions use the nimble.directory install. This
currently downloads from https://notabug.org/vktec/nimbox. The code
there is out of date, containing a type error bug which has been
fixed in this repository. Until the nimble.directory is updated to
point to this repository, this updates the readme to reflect the
manual download and install method rather than the automatic.

closes #1